+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Typed connect views [@fringer2423](https://github.com/fringer2423)
 - Typed QRCode views
 - Typed rate info views
+- Typed rate views
 
 ### Изменено